backend/requirements-base.txt
# Senior MHealth - 기본 의존성
# Core dependencies for all backend services

# === Core Framework ===
fastapi==0.115.12
uvicorn[standard]==0.32.1
pydantic[email]==2.10.4
python-dotenv==1.0.1
python-multipart==0.0.20

# === Google Cloud Services ===
google-cloud-speech==2.23.0
google-cloud-storage==2.16.0
google-cloud-firestore==2.18.1
firebase-admin==6.1.0

# === Audio Processing ===
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
audioread==3.0.1

# === ML and Data Processing ===
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
scipy==1.11.4

# === HTTP and Async ===
httpx==0.28.1
aiohttp==3.9.3
aiofiles==23.2.1

# === Monitoring & Logging ===
prometheus-client==0.20.0
structlog==24.1.0
python-json-logger==2.0.7

# === Utilities ===
python-dateutil==2.8.2
pytz==2024.1
pyyaml==6.0.1